Add second part of parser.
Uses mutual recursion, with parseExpression called first.
The reason parseExpression is called first is that this is not a lisp.
Programs in Egg will not begin with a function application.
They begin with a word, like do() or some other special form.
In Egg, the parens come *after* the function name, as in non-lisp languages, so whether a word is a function is only detectable then.